fix pep8

+++ yatube_api/api/views.py
@@ -1,4 +1,3 @@
-from django.shortcuts import render
 from rest_framework import viewsets, permissions
 from rest_framework.decorators import action
 from rest_framework.response import Response
@@ -14,7 +13,7 @@ class PostViewSet(viewsets.ModelViewSet):
     permission_classes = [IsAuthorOrReadOnly, permissions.IsAuthenticated]
 
     def perform_create(self, serializer):
-        serializer.save(author=self.request.user) 
+        serializer.save(author=self.request.user)
 
     @action(detail=True, methods=['get', 'post'])
     def comments(self, request, pk=None):
@@ -27,13 +26,16 @@ class PostViewSet(viewsets.ModelViewSet):
             serializer = CommentSerializer(data=request.data)
             if serializer.is_valid():
                 serializer.save(author=request.user, post=post)
-                return Response(serializer.data, status=status.HTTP_201_CREATED)
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+                return Response(serializer.data,
+                                status=status.HTTP_201_CREATED)
+            return Response(serializer.errors,
+                            status=status.HTTP_400_BAD_REQUEST)
 
     @action(detail=True,
             methods=['get', 'put', 'patch', 'delete'],
             url_path='comments/(?P<comment_pk>[^/.]+)',
-            permission_classes=[IsAuthorOrReadOnly, permissions.IsAuthenticated])
+            permission_classes=[IsAuthorOrReadOnly,
+                                permissions.IsAuthenticated])
     def comment_detail(self, request, pk=None, comment_pk=None):
         post = self.get_object()
         try:
@@ -41,19 +43,23 @@ class PostViewSet(viewsets.ModelViewSet):
         except Comment.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
 
+        self.check_object_permissions(request, comment)
+
         if request.method == 'GET':
             serializer = CommentSerializer(comment)
             return Response(serializer.data)
         elif request.method in ['PUT', 'PATCH']:
-            serializer = CommentSerializer(comment, data=request.data, partial=request.method == 'PATCH')
+            serializer = CommentSerializer(comment,
+                                           data=request.data,
+                                           partial=request.method == 'PATCH')
             if serializer.is_valid():
                 serializer.save()
                 return Response(serializer.data)
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            return Response(serializer.errors,
+                            status=status.HTTP_400_BAD_REQUEST)
         elif request.method == 'DELETE':
             comment.delete()
             return Response(status=status.HTTP_204_NO_CONTENT)
-
 
 
 class GroupViewSet(viewsets.ReadOnlyModelViewSet):
@@ -73,4 +79,4 @@ class CommentViewSet(viewsets.ModelViewSet):
         post_id = self.kwargs.get("post_id")
         # И отбираем только нужные комментарии
         new_queryset = Comment.objects.filter(post=post_id)
-        return new_queryset     
+        return new_queryset
